backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
alembic==1.14.1
gunicorn==23.0.0
redis==5.2.1
marshmallow==3.26.1
flask-cors==5.0.0
flask-limiter==3.12
flask-restx==1.3.0
